+++ src/index.js
@@ -16,8 +16,13 @@ import cleanIcon from "./../assets/icons/clean.svg";
 
 import "../assets/index.styl";
 
+const Icons = Quill.import("ui/icons");
+Object.assign(Icons, { clean: cleanIcon });
+
 Quill.register(
   {
+    "ui/icons": Icons,
+
     "modules/wordCount": wordCount,
     "modules/imageResize": ImageResize,
     "modules/image": Image,
@@ -37,25 +42,10 @@ class QlQuill {
     // 阻止谷歌翻译输入框
     this.container.classList.add("notranslate");
 
-    this.replaceIcon();
     this.instantiateEditor(options);
     this.setEditorContents(options.value);
 
     this.setContent = this.setEditorContents.bind(this);
-  }
-
-  // 替换图标
-  replaceIcon() {
-    let Icons = Quill.import("ui/icons");
-    Icons = Object.assign(Icons, {
-      clean: cleanIcon,
-    });
-    Quill.register(
-      {
-        "ui/icons": Icons,
-      },
-      true
-    );
   }
 
   instantiateToolbar(options) {
@@ -85,7 +75,7 @@ class QlQuill {
 
     const toolbar = this.instantiateToolbar(options);
 
-    let editorOption = {
+    let option = {
       theme: "snow",
       modules: {
         toolbar: {
@@ -154,33 +144,34 @@ class QlQuill {
     };
 
     if (options.limit) {
-      options.limit = Number(options.limit) || 1000;
-      editorOption.modules.wordCount.limit = options.limit;
+      option.modules.wordCount = {
+        limit: Number(options.limit) || 1000,
+      };
     }
 
     if (options.imageResize) {
       if (typeof options.imageResize === "boolean") {
-        editorOption.modules.imageResize.modules = editorOption.modules.imageResize.modules.concat(
+        option.modules.imageResize.modules = option.modules.imageResize.modules.concat(
           ["Resize", "DisplaySize"]
         );
       } else if (typeof options.imageResize === "object") {
-        editorOption.modules.imageResize = imageResize;
+        option.modules.imageResize = imageResize;
       }
     }
 
     if (options.image) {
       if (typeof options.image === "function") {
-        editorOption.modules.toolbar.handlers.image = () =>
+        option.modules.toolbar.handlers.image = () =>
           options.image(this.insertImage);
       } else if (typeof options.image === "object") {
         if (options.image.action) {
-          editorOption.modules.toolbar.handlers.image = () =>
+          option.modules.toolbar.handlers.image = () =>
             this.uploadImages(options);
         }
       }
     }
 
-    this.editor = new Quill(this.container, editorOption);
+    this.editor = new Quill(this.container, option);
 
     this.editor.on(
       "editor-change",
@@ -294,8 +285,8 @@ class QlQuill {
     }
     const index = selection.index || 0;
     this.editor.insertEmbed(index, "image", {
-      url: src,
-      latex,
+      src,
+      "data-latex": latex,
     });
     this.editor.setSelection(index + 1);
   };