+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - editor
   - 字数统计 使用 modules 形式实现
   - 修复空格不显示问题
+  - iamge 标签保留 width、height 等属性
 - webpack 配置
   - 新增 raw-loader 读取 svg 文件源文本